Desktop/7940Template/requirements.txt
fastapi~=0.110.0
uvicorn
pydantic
pytest
httpx
pyyaml~=6.0.1
python-telegram-bot==13.7
telegram~=0.0.1
requests~=2.31.0
psycopg2-binary